bbs: 接口调整

+++ bbs/bbs-manager/src/main/java/com/github/kop/bbs/service/impl/ScoreRoleServiceImpl.java
@@ -103,7 +103,7 @@ public class ScoreRoleServiceImpl implements ScoreRoleService {
         Long sumScore = scoreRecordService.sumScoreByRole(scoreRole.getRoleId(), userId, startTime,
             endTime);
 
-        Integer maxScore = scoreRole.getMaxScore();
+        Long maxScore = scoreRole.getMaxScore();
         if (maxScore == null || maxScore == 0) {
             return scoreRole;
         } else {
